Remove shouldComponentUpdate and make the component Pure
In the case of a fully controlled component, componentWillReceiveProps was
getting called and updating the state, and then shouldComponentUpdate was
getting called, comparing the new nextProps.value to the ALREADY UPDATED
this.state.rrule and returning false, meaning the component was never
re-rendered. As a quick fix, removing shouldComponentUpdate entirely and
marking the component as Pure to allow React to optimize for us.